refactor(admin): extract InstanceConfig and FundingSupporters
Configuration (general/federation/branding/oauth/webrtc tabs) and the
Funding & Supporters section move into self-contained components with
their own state, methods and styles. AdminPanel.vue is now 3.2k lines,
down from 7.5k.
